requirements/requirements.txt
mycroft-messagebus-client~=0.9,!=0.9.2,!=0.9.3
requests
requests_cache~=0.6
beautifulsoup4~=4.9.3
lxml~=4.5.2
filelock~=3.2.0
pendulum~=2.1.2
geopy~=2.1.0
timezonefinder~=5.2.0
netifaces~=0.10.9
nltk~=3.5
ruamel.yaml~=0.16.12
pydub~=0.25.1
fastlang==0.0.4
lingua-nostra~=0.4.5
langdetect~=1.0.8
translate
boto3
ovos_utils>=0.0.6
neon_mq_connector~=0.1.2
combo-lock~=0.1
soundfile~=0.10.3
librosa~=0.8
filelock~=3.2.0